createCard
change Card Entity to include methods that set values for createCard mutation

add create Card to Pipe Class.

create the gql file.

There is a lot of things to do, like set the data types that make easy transform
and validate data to the entity.

implement schema validation.

and others.