create cache that is never clear
Cache::ALLWAYSVALIDCACHE = -1
when you need that a file always exist because you need then in some special case.
like serial copies to know the data changes.
Schema is one of this special cases
even if it is old, The system need it as reference to build the entities data\types

ALLWAYSVALIDCACHE permit never clear the cache file and PIpefy::useCase(true) make
the system try to consume cache before query.