move found() to loaded()
it express the intention better.
this method is to know if the entity  was created from query data result.

Being loaded is the only way to have __get() data response.
When you going to create a new entity you need a not loaded entity to
write it.
but if you try to read data from a unloaded entity you gonna throw an
Exception.