Schema Information update
Now is possible to retrieve the schema of a Pipe and a Field.
It is used by the system to hydrate data fields with their rules and
make possible validate it.

By the schema The system know the data type, if it is a deleted field, if is required
or not, the id and others.

This is usefull to grant consistency when  setField in a card(and in a future table record
implementation).

Schema\Data\Type
- collectionOf : used to array fields, can define the colletable data type.

Schema\Data\Type
- ShortText (maxlength 255 as tested at pipefy)
- LongText (maxlength not limited as tested at pipefy)